+++ dashboard/api/src/main/java/com/dashboard/api/Repository/DashBoardRepository.java
@@ -3,5 +3,5 @@ package com.dashboard.api.Repository;
 import org.springframework.data.repository.CrudRepository;
 import com.dashboard.api.Entity.DashBoard;
 
-public interface DashBoardRepository extends CrudRepository<DashBoard, Integer> {
+public interface DashBoardRepository extends CrudRepository<DashBoard, Long> {
 }
